first refactor set
