--- examples.http
@@ -1,15 +1,0 @@
-@token = 783830259:AAG39d-ML_LEM1WR4Ehi1n5GHmzPUYP4YyY
-
-GET https://api.telegram.org/bot{{token}}/getUpdates
-
-###
-
-POST https://api.telegram.org/bot{{token}}/sendmessage?chat_id=106258322&text=Hellj
-
-###
-
-POST https://api.telegram.org/bot{{token}}/sendAnimation?chat_id=106258322&animation=CgADBAADDaAAAvwYZAcBcVMvYX2ftQI HTTP/1.1
-
-###
-
-POST https://api.telegram.org/bot{{token}}/sendSticker?chat_id=106258322&sticker=CgADBAADDaAAAvwYZAcBcVMvYX2ftQI
