--- strings.yaml
@@ -1,1 +1,0 @@
-wetaher_string:
